uninstall mocha; install jest; rename some files conflicting with jest conventions
